Adding a modify function
Step 8: The modify function uses a HttpPut request type. When you send a HttpPut request, two things will be required:

The index position of the item being modify in the context (Ex. https://localhost:5001/api/todo/1 would be the long id=1)

The json object passing with the HttpPut request (Ex.   {
    "Id":1,
    "Name":"feed fish",
    "IsComplete":true
  })

Note that the input is an JSON object and it must match TodoItem variable type.