Call the web api with jQuery
Step 10: An addtional piece for if you want to use jQuery to interact with the API.

The functions in site.js uses ajax as a wrapper to the TodoController.cs Http request functions.

index.html is hooked up to site.js. jQuery Ajax calls from site.js are wrapper functions that makes the requests to the TodoController api. The functions can also update parts of the webpage without having to refresh the entire page.

Learning jQuery and Ajax was intitially confusing. I was confused initially what jQuery is, and how Ajax is related to jQuery.

JQuery is a javascript framework that reduces the complexity of using javascript on a website. It makes it easier to interact with the DOM (A Document Object Model which treats xml, html, and xhtml as a tree structure) by creating lots of high level functionality that can be used to search and interact with the DOM.

Ajax is a subsection of javascript, a technique that does client-side browser communication with the server without having to refersh the page.

jQuery Ajax implements Ajax's high level interface abstractly such that the developer is shield from the complexity of multi-browser support in making the requests.

https://stackoverflow.com/questions/931908/what-is-the-difference-between-ajax-and-jquery-and-which-one-is-better